+++ UpgradeWorld/commands/ObjectsList.cs
@@ -6,7 +6,7 @@ public class ObjectsListCommand {
       return DataParameters.Parameters;
     }, DataParameters.GetAutoComplete());
     new Terminal.ConsoleCommand("objects_list", "[id1,id2,...] [...args] - Lists objects showing their position and biome.", (args) => {
-      DataParameters pars = new(args, true);
+      DataParameters pars = new(args, false);
       if (!pars.Valid(args.Context)) return;
       if (Helper.IsClient(args)) return;
       new ListObjectPositions(args.Context, pars.Ids, pars);