+++ UpgradeWorld/operations/objects/ListObjectPositions.cs
@@ -4,7 +4,7 @@ using Service;
 namespace UpgradeWorld;
 /// <summary>Lists positon and biome of each entity.</summary>
 public class ListObjectPositions : EntityOperation {
-  public ListObjectPositions(Terminal context, IEnumerable<string> ids, DataParameters args) : base(context) {
+  public ListObjectPositions(Terminal context, List<string> ids, DataParameters args) : base(context) {
     ListPositions(ids, args);
   }
   private string GetData(ZDO zdo, List<string> prints) {
@@ -15,9 +15,11 @@ public class ListObjectPositions : EntityOperation {
       return DataHelper.GetData(zdo, split[0], type);
     }).ToList());
   }
-  private void ListPositions(IEnumerable<string> ids, DataParameters args) {
+  private void ListPositions(List<string> ids, DataParameters args) {
+    if (ids.Count == 0) ids.Add("*");
+    var prefabs = ids.SelectMany(GetPrefabs).ToHashSet();
     var zdos = GetZDOs(args);
-    var texts = ids.Select(id => {
+    var texts = prefabs.Select(id => {
       return GetZDOs(zdos, id).Select(zdo => id + " " + zdo.m_uid.ID + " " + Helper.PrintVectorXZY(zdo.GetPosition()) + ": " + GetData(zdo, args.Prints));
     }).Aggregate((acc, list) => acc.Concat(list));
     if (args.Log) Log(texts);